adding watchdog timer to rx fifos in ethernet firmware to prevent out of sequence command handling when protocol errors occur

+++ scripts/shiftToStandAlone.py
@@ -11,7 +11,7 @@ import os #for isdir abspath dirname system
 
 print
 print "***********************\n"
-print "Look for 'Success!' at end of print out.\n" + \
+print "Look for 'Done!' at end of print out when script completes.\n" + \
 	"Setting up and Installing...\n"
 print
 
@@ -30,8 +30,17 @@ print args
 print
 print
 
+if os.path.isfile(args.source) == False:
+	print 'Error! Source file does not exist!'
+	exit()
+
 print 'copying source to dest'
 os.system("cp " + args.source + " " + args.dest)
+
+if os.path.isfile(args.dest) == False:
+	print 'Error! Copy failed to destination file!'
+	exit()
+
 print 'if vivado, changing add_files to import_files'
 os.system("sed -i s/add\_files/import_files/g " + args.dest)
 print 'if ise, changing xfile add to xfile add -copy'
@@ -40,7 +49,7 @@ os.system("sed -i /xfile\ add/\ s/\$/\ \-copy/g " + args.dest)
 	
 print
 print "***********************\n"
-print 'Success!'
+print 'Done!'
 print
 print
